+++ src/org/alixia/chatroom/Launch.java
@@ -1,6 +1,10 @@
 package org.alixia.chatroom;
 
+import org.alixia.chatroom.impl.commands.Commands;
+import org.alixia.chatroom.texts.BoldText;
+
 import javafx.application.Application;
+import javafx.scene.paint.Color;
 import javafx.stage.Stage;
 
 public class Launch extends Application {
@@ -9,7 +13,16 @@ public class Launch extends Application {
 
 	@Override
 	public void start(Stage primaryStage) throws Exception {
-		new ChatRoom(primaryStage);
+		new ChatRoom();
+		ChatRoom.INSTANCE.setStage(primaryStage);
+
+		ChatRoom.INSTANCE.println("Setting up commands...", Color.BISQUE);
+
+		ChatRoom.INSTANCE.print("Commands took ", ChatRoom.SUCCESS_COLOR);
+		long initialTime = System.currentTimeMillis();
+		new BoldText(Commands.getTime() - initialTime + " ", Color.FIREBRICK).print(ChatRoom.INSTANCE.console);
+		ChatRoom.INSTANCE.println("milliseconds to load!", ChatRoom.SUCCESS_COLOR);
+
 	}
 
 	public static void main(String[] args) {
